split out logindialog
