+++ src/LoginDialog.tsx
@@ -3,8 +3,9 @@ import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 import Modal from 'react-responsive-modal';
 import { useFormik } from 'formik';
-import { login } from './service';
+import { login, loadAuthenticationOptions } from './service';
 import { useCustomerData, useTranslation } from './app-state';
+import { PasswordLoginForm }from './PasswordLoginForm';
 import { createRegistrationUrl } from './routes';
 import { ReactComponent as CloseIcon } from './images/icons/ic_close.svg';
 
@@ -22,57 +23,15 @@ interface FormValues {
 
 export const LoginDialog: React.FC<AppModalLoginMainProps> = (props) => {
   const { handleModalClose, openModal } = props;
-  const { setCustomerData } = useCustomerData();
+  
+  
+  const authenticationOptions = loadAuthenticationOptions();
   const { t } = useTranslation();
-  const registrationUrl = createRegistrationUrl();
 
-  const [failedLogin, setFailedLogin] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
-
-  const initialValues:FormValues = {
-    emailField: '',
-    passwordField: '',
-  };
-
-  const validate = (values:any) => {
-    const errors:any = {};
-    if (!values.emailField) {
-      errors.emailField = t('required');
-    }
-    if (!values.passwordField) {
-      errors.passwordField = t('required');
-    }
-
-    return errors;
-  }
-
-  const {handleSubmit, handleChange, resetForm, values, errors} = useFormik({
-    initialValues,
-    validate,
-    onSubmit: (values) => {
-      setIsLoading(true);
-      login(values.emailField.toLowerCase(), values.passwordField)
-        .then((result) => {
-          handleModalClose();
-          setIsLoading(false);
-          setCustomerData(result.token, result.customer_id);
-        })
-        .catch(error => {
-          setIsLoading(false);
-          setFailedLogin(true);
-          console.error(error);
-        });
-    },
-  });
-
-  const registerNewUser = () => {
-    handleModalClose();
-  }
-
+  
   const handleClose = () => {
-    setFailedLogin(false);
     handleModalClose();
-    resetForm();
   }
 
   return (
@@ -89,40 +48,15 @@ export const LoginDialog: React.FC<AppModalLoginMainProps> = (props) => {
             <CloseIcon />
           </button>
         </div>
+        
+        <PasswordLoginForm handleModalClose={handleModalClose} isLoading={isLoading} setIsLoading={setIsLoading} />
+        
+        {// TODO:
+          // Map through all the different login profiles here...
+        }
 
-        <div className="logindialog__body">
-          <div className="logindialog__feedback">
-            {failedLogin ? t('invalid-email-or-password') : ('')}
-          </div>
-          <form className="epform" id="login_modal_form" onSubmit={handleSubmit}>
-            <div className={`epform__group ${errors.emailField ? '--error' : ''}`}>
-              <label className="epform__label" htmlFor="emailField">
-                {t('email')}:
-              </label>
-              <input className="epform__input" id="emailField" type="text" onChange={handleChange} value={values.emailField} />
-              <div className="epform__error">
-                {errors.emailField ? errors.emailField : null}
-              </div>
-            </div>
-            <div className={`epform__group ${errors.passwordField ? '--error' : ''}`}>
-              <label className="epform__label" htmlFor="passwordField">
-                {t('password')}:
-              </label>
-              <input className="epform__input" id="passwordField" type="password" onChange={handleChange} value={values.passwordField} />
-              <div className="epform__error">
-                {errors.passwordField ? errors.passwordField : null}
-              </div>
-            </div>
-            <div className="epform__group --btn-container">
-              <button className="epbtn --primary" id="login_modal_login_button" type="submit" disabled={isLoading}>
-                {t('login')}
-              </button>
-              <Link to={registrationUrl} className="epbtn --primary" id="login_modal_register_button" onClick={registerNewUser}>
-                {t('register')}
-              </Link>
-            </div>
-          </form>
-        </div>
+        
+
       </div>
     </Modal>
   );
