+++ src/service.ts
@@ -124,6 +124,36 @@ export interface CustomerToken {
   expires: any;
 }
 
+// HAX - For now this can just be a non async function...
+export function loadAuthenticationOptions(): object {
+  // TODO: return the options in the authentication options...
+  return {
+    "data": {
+      "type": "customer-authentication-settings",
+      "allow_password_authentication": true,
+      "links": {
+        "self": "https://api.moltin.com/v2/customer-authentication-settings"
+      },
+      "relationships": {
+          "authentication-realms": {
+              "data": [
+                  {
+                      "type": "authentication-realm",
+                      "id": "369ad4a4-ee67-48b0-x347-t50a6e61d83d",
+                      "links": [
+                          {"self": "https://api.moltin.com/v2/authentication-realm/{id}"},
+                      ],
+                  }
+                ]
+          }
+      },
+      "meta": {
+          "clientId": "steveisthebest"
+      }
+    }
+  }
+}
+
 export async function loadEnabledCurrencies(): Promise<Currency[]> {
   const moltin = MoltinGateway({ client_id: config.clientId });
   const response = await moltin.Currencies.All();
